requirements.txt
# Core Framework Dependencies
mcp==1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Agentic Framework Dependencies (Phase 1.2)
crewai>=0.1.0
langgraph>=0.1.0
autogen-agentchat>=0.2.0

# LLM and API Dependencies
openai>=1.0.0
anthropic>=0.8.0
langchain>=0.1.0
langsmith>=0.1.0

# Infrastructure Dependencies  
asyncio
aiohttp>=3.8.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
redis>=4.5.0
celery>=5.3.0

# Monitoring and Logging
structlog>=23.1.0

# Development Dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
flake8>=6.0.0

# Utility Dependencies
dataclasses-json>=0.6.0
typing-extensions>=4.0.0
requests>=2.31.0